(PUP-7412) Puppet device, ability to pass single certname. (#5768)
Prior to this commit `puppet device` would process all devices configured in `device.conf`. This behaviour is undesirable as users need the ability to test and run against specific nodes.

This commit will allow a user to pass a single certname to `puppet device`. For example, if there are four devices in `device.conf` and we only want to run against a single host, `puppet device --target host1` will only fetch and apply the configuration for `host1`.